Fix Environments ----------------
A host's environment is now a first class object and stored in the
nodes table rather than being a host attribute.  But is still
present as an intrinsic attribute.  By default hosts are not part
of any environment.

WARNING: SCHEMA CHANGED

To use this commit do the following on your Frontend with the older
schema:

alter table nodes add column Environment varchar(64) default NULL;

This will add the required column.  You will know this is correct when
'stack list host' shows the environment is the output.