Post Jury Duty sync
This code works with the same commit on the ubuntu pallet with the
above heading.

More cleanup of dead code (Solaris) and made the XML parsing more
generic for RedHat so we don't hove to list out the tags in the
NodeFilters.  This will make it easier to add (or remove) support
for other tags in the future.

Split out the "debug" and "order" sections into what are not called
<profile> <chapter>s (were called <section)s which was confusing.

Still need to update the client side to understand the new format
of the kickstart profile.