stack run pallet now handle the <boot> section and non-bash shell <post> sections.
This is not fully tested, since I don't want to 'stack run pallet'
on my build host.  But it does look right :)

The key code is all in pylib/stack/redhat/gen.py in case this
causes problems for anyone.  Just want to get this into the
tree right away so others can start do a barnacle install again.