Added Groups - READ THE FULL LOGMESSAGE ---------------------------------------
A group is a set of host with absolutely no semantics, just a set of hosts.

New commands:
	stack [add|remove|list] group
	stack [add|remove|list| host group

Removed the "membership" appliance confusion, what we used to call a
membership is now just the appliance long name. This has effectively
been true for several releases so this just makes the interface less
confusing to the user with no real functional change. BUT this means
the DB SCHEMA HAS CHANGED, and lots of dead code has been removed.

The word "membership" is now re-used in the NEW DB SCHEMA to represent
the mapping for hosts to groups, but is not exposed on the command line.

Everthing on the stacki pallet has been updated for this but the stacki-pro
and stacki-ux pallets will likely need updating as well.

TODO:
	add groups to the HostArgumentProcessor
	allow groups to be specified on the host spreadsheet
	should there be intrinsic attributes for this? (e.g. group.name=[true|false])

Sounds Awesome How can I Use This?
----------------------------------

It is awesome, but the DB has changed so you need to update your DB schema.

1)

> alter table appliances change Membership LongName varchar(64);

Also look in database-schema.xml and add the tables for groups & memberships

2)

Rebuild and install the stack-command package