add optional 'authorized_keys' attribute
Allows for copying one or more ssh public keys to installing nodes.

To use, set the value to one of:
* the contents of a single public key, OR
* a URL (http(s)://, ftp://, file://) which contains one or more keys

If using a URL, it must be accessible to the frontend if pxebooting, or the installing node if booting from dvd/flash.
The content of the URL may be one or more public keys, one per line in the format of the `~/.ssh/authorized_keys` file.