cond.py bandaid
The real issue is we have the pallets and carts attributes that are
represented as Python lists. This means the old code that
correctly handled '.' was changed because the attribute is now a list.

This is the core issue, for now the cond.py code looks for the case
and handles it, but this opens the door for other types of attributes.

Long term everything needs to be a string, BUT we have several node
files that are using the pallets attribute. This needs to remain
supported until we stop using the pallets attribute.

I feel dirty.