report host interface - change dhcp code
If an interface is 'default' and has the 'dhcp' option set the following:
   DHCLIENT_SET_HOSTNAME="yes"
   DHCLIENT_SET_DEFAULT_ROUTE="yes"
For non-default dhcp interfaces set the above to "no"

This was needed for AWS but will still work in Vagrant.

Also cleaned up the code to use  'list host interface expanded=true' so
we could remove the call to 'list network'.