RedHat
Use Kickstart_PrivateAddress as the tracker, this is what SLES does.
But we need to create stack_site twice
     initrd (new code)
     %pre (storage.xml)

Previously we used the next-server dhcp option which is fine for
bare metal but doesn't work in AWS.

So yeah, this is an AWS fix that happens to make RedHat and SLES
do the same thing.

Cloud rules !!!!