Change how RPM.STRATEGY=copy works.
When using RPM.STRATEGY=copy, the expectation is that all
RPM files are already part of the source tree. However,
if RPMS are downloaded during build time, the copy targets
are created before any RPMS are downloaded, causing the copy
command to fail.

Instead, we now source makefile target directly, so that the
copy target isn't calculated until make is called directly