Use rpm to infer name,arch,version, release :fire:
BUG: When Stacki moved to python3, the `import rpm` and related
python calls failed. As a fallback, we relied on parsing the
filename to get basename architecture, version, and release,
using `string.split` and regex-es.
This meant that if an RPM has a non standard file name, we
don't get the RPM information, and ignore the file.

FIX: Use the `rpm` utility to get all the information that we need.

FIXES: JIRA STACKI-323