Fix ludicrous client exeptions that were crashing the client
If the network changed while in the middle of package
installation, requests would timeout and raise an exeption
that was unhandled.

All execption are now handled with logs about which function
is throwing the exception.

Also, every request is retried 3 times to give time for
the new connection to come up if something did happen
with the network. Letting requests handle this wouldn't
create a new connection.