Stacki SLES12sp3 ISO is bootable
Wizard for FE install is not in place yet

Jumbo os Stacki/SLES will be needed w/o this the Stacki pallet
is bootable but there won't be any OS package only the installer
bit are on the media.

Add stupid respin script to feature/bootable, remember to nuke this
before merging back onto develop.