When using shortnames, stack run host can break
When an interface in the database is set to use shortnames, that
becomes its quasi-default interface, since that shortname is how
run host connects to the node.

However, it's possible that this interface is not on a network that
the frontend can access. If that's the case, the frontend cannot run
a run host command against it.

To avoid this problem, we use the 'stack.network' attribute. This will
tell 'run host' which network to use when talking to hosts. If this
attribute doesn't exist, 'run host' will use the shortname
(default old-style behaviour)

Use fully qualified domain names when syncing firewall

Use getRunHosts method to get fqdn for  hosts. Scaling improvements

Decode the output of subprocess before returning
Fix report host network, and make is OS specific
Fix sync host network again to run the entire command in parallel