FEATURE: Make table output format more consistent.
`stack list` commands with NULL output in the last column will now have -'s printed
up to the width of that column's header as in other columns.

This allows cell filling in the 'channel' column of 'stack list host interface',
without cluttering the output with lots of -'s in 'stack list host' if 'comment' is long.