FEATURE: Add user settable host metadata
This is similar to the AWS meta-data, the idea being a container for
arbitrary data unique to a host. `stack set host metadata` is used
to set this and the data itself is made available from a read-only
host attribute called `metadata`.

The same thing could have been accomplished using a standard attribute,
but the intention is to have the semantics that this is unique to a host
and not subject to the standard attribute inheritance rules.

BREAKING CHANGE: schema addition

```
$ mysql cluster
MariaDB [cluster]> alter table nodes add column MetaData text default NULL;
```