FEATURE: turn status field into a real feature
Enhanced the message queue status reporting to be hostid based (not
hostname) and added a ttl to all messages. The first part fixes issues
when hosts are renamed and the status gets lost, now everything is
keyed off the ID from the nodes table (permanent for lifetime of
host). The second allows us the age out messages at different rates.

Instrumented multiple parts of the installer to send more fine grained
status messages. The typical workflow of a system install will see the
following messages:

| message                   | ttl  |
| ------------------------- | ---- |
| install profile request   |   30 |
| install profile sent      |   30 |
| install profile           |  300 |
| install download          | 3600 |
| install stage=pre         | 3600 |
| install stage=pre-package | 3600 |
| install download          | 3600 |
| install stage=post        | 3600 |
| install reboot            |  300 |
| install stage=boot-pre    | 3600 |
| install stage=boot-post   | 3600 |
| online                    |  120 |

The `install download` is seen both before the second stage image
downloads and rpm tracker lookups. All `install *` messages should be
thought of as the previous `install` and the extra information is for
debugging and may change at any time.

The `online` state was previously called `up` and just meant that the
installer's first pass was done and the host was on the network. For
SuSe it could be seens even though the second pass of the installer
was active. To improve on this Stacki now tracks is `sshd` is running
as a seperate state (piggy-back on the existing health message). The
inspired a new command `stack list host status` which a `state` and
`ssh` fields. The `state` field is the previous health message as
described above. The `ssh` field is `up` IFF sshd is running, and None
if it is not (TTL same as the `state` health).

The new `stack list host status` command means:

- `stack list host` `status` field is now *deprecated* and will be
  removed in the next release. Just still present to keep existing
  consumers from crashing but it will always says *deprecated*.

- pluggins can be used to add move status (is service X runnning?)

Other stuff:

        * update Redis to 4.0.11
	* change rmq to smq
	* health channel is now json payload (backwards compat with
          old text - but will deprecate)
	* stack-mq is now in the installer (SLES only)
	* stack-publish.py is gone replaced by updated smq-publish
	* fixed some leftover Python3 transition encode()/decode() bugs
	* remove old processors related to StackIQ Enterprise
	* changed `message` field in `Message` to `payload` (less confusing)
	* added psutil to foundation-python-packages
	  * this requires refreshing version of all python code
	  * hard code PyMySQL to 0.8.1 (newer version require newer pip)

ToDo:

	* Instrument RedHat installer
	* Need a `install start` state (where to trigger it?)
	* fix test_enable_discover.py