BUGFIX: stack list host unreliable status field
Enhanced the message queue status reporting to be hostid based (not
hostname) and added a ttl to all messages. The first part fixes issues
when hosts are renamed and the status gets lost, now everything is
keyed off the ID from the nodes table (permanent for lifetime of
host). The second allows us the age out messages at different rates.

Instrumented multiple parts of the installer to send more fine grained
status messages. The typical workflow of a system install will see the
following messages:

| message                   | ttl  |
| ------------------------- | ---- |
| install profile request   |   30 |
| install profile sent      |   30 |
| install profile           |  300 |
| install download          | 3600 |
| install stage=pre         | 3600 |
| install stage=pre-package | 3600 |
| install download          | 3600 |
| install stage=post        | 3600 |
| install reboot            |  300 |
| install stage=boot-pre    | 3600 |
| install stage=boot-post   | 3600 |
| up                        |  120 |

The `install download` is seen for before the second stage images and
rpm tracker lookups (hence shows up twice). All `install *` messages
should be thought of as the previous `install` and the extra information is
for debugging and may change at any time.

Other stuff:

	* change rmq to smq
	* stack-mq is not in the installer (SLES only)
	* stack-publish.py is gone replaced by updated smq-publish
	* fixed some leftover Python3 transition encode()/decode() bugs
	* remove old processors related to StackIQ Enterprise
	* changed `message` field in `Message` to `payload` (less confusing)

ToDo:

	* Instrument RedHat installer
	* Need a `install start` state (where to trigger it?)