BUGFIX: Update interface information with MAC addresses correctly
When installing a backend node, all the interfaces are reported
back by the installer. Any interfaces missing on the frontend,
will be added to the database.

The Bug is - when interfaces are specified in the database, with
all the information except the MAC address, this call fails to
update the MAC address.

This commit fixes that.

FIXES: JIRA STACKI-585