INTERNAL: SQL fixes for add network and add pallet
SQL fixes and lots of tests for:

* add network
* add pallet

FEATURE: Added two new pytest fixtures to test-framework

*run_file_server* - This will run a simple http server on 127.0.0.1:8000
serving up the files in /export/test-files

*rmtree* - Add this fixture and then call it in your code to simulate
removing a directory and all its files. After the test finishes, the directories
are restored.

FEATURE: New count() method on DatabaseConnection

The new count() method in the DatabaseConnection class will return
a count of the number of matching items in the database.

The command query should start with the column in parentheses you
wish to count.

The return value will either be an int or None if something
unexpected happened.

Example: `count('(ID) from subnets where name=%s', (name,))`

BUGFIX: Adding a pallet that was already mounted was broken