FEATURE: Tags (sort of like Attributes but not)
This is the stat of a Tags feature. Tags can be applied to various objects in the
Stacki Database but do not roll up into XML entities when building host profiles.
Tags are only visible through the CLI/API and for now only with the following
commands:

- `stack add pallet tag`
- `stack set pallet tag`
- `stack list pallet tag`
- `stack remove pallet tag`

BREAKING CHANGE: schema addition

```
$ mysql cluster
CREATE TABLE tags (
        Scope           enum ('box', 'cart', 'host', 'network', 'pallet'),
        Tag             varchar(128) NOT NULL,
        Value           text,
        ScopeID         int(11),
        INDEX (Scope),
        INDEX (Tag),
        INDEX (ScopeID)
);
```