BUGFIX: Change how getBoxPallet returns information
This branch started as a bugfix branch for "run pallet".
"run pallet" did not work if there were 2 pallets of the same
name on the frontend, but only one was enabled for the frontend box.

This required a re-architecting of the "run pallet" command, which
involved changing how getBoxPallet function returned information.

With this change "getBoxPallets" function returns output in the
same format as "getPallets" from the PalletArgument Processor.
This makes it easy to compare the 2 data structures.

This also means that, all commands using getBoxPallets need to change.
As part of this cleanup, "report host repo" was also cleaned up.