BUGFIX: The `output-network` of a firewall rule should generate a `-o` flag
BUGFIX: `report host interface` for redhat was incorrectly checking for the existance of an interface name, not the network.

BUGFIX: CNAMEs in `report zones` were broken because the code was using the IP address, not the device, to look up entries in the aliases table.

INTERNAL: SQL fixes, cleanups, and tests for:
* report dhcpd
* report host firewall
* report host interface
* report zones

INTERNAL: New test fixtures to fake out the OS detection in Stacki