FEATURE: Better Shadow Attributes
Split the shadow attributes out of the `cluster` database and into a
their own `shadow` database. This is the start of a locked down secrets
database where host keys and other things can be stored.

BREAKING CHANGE:

Added a whole new database and redid permissions. This is a scorched earth
commit, with no path forward for existing Frontends.

INTERNAL: Added `--bootstrap` flag to pip2src