BUGFIX: Cache clearing had incorrect check
The previous code was just looking for `select` in the query, but this was falsely preventing cache clearing for INSERT and UPDATE queries with SELECT subqueries. Now a query needs to start with `select` (case-insensitive) to prevent the cache from getting cleared.

INTERNAL: Lowered the time to run all integration tests by changing the mysqldump flags and increasing the frontend RAM to 2GB.

I tried lots of other things, such as storing MySQL's data files in a ramfs and only reverting the database is there was a non-select statement run, but it had little to no impact on runtime speed.

I was also able to simplify the database revert code, because we no longer need to support Stacki build MySQL.