INTERNAL: Remove kdump configuration from Stacki
Stacki currently lays down a default kdump configuration.
This configuration comes from our initial collaboration with TD,
and reflects how kdump should be configured for a TD Database machine.

However, this is then zero-ed out by the tdc-infrastructure pallet
just so that an OS package can then lay down another configuration
on top of it.

Since this is not used for anything outside of TD, nuke it