FEATURE: Add `--use-src` flag to integration test set-up.sh
The new `--use-src` flag for `set-up.sh` will cause the local Stacki source to be mounted inside the VM that is created. The ISO provided files are then removed and replaced with symlinks to the local source files. This allows you to run test against the local source changes, or to login via `vagrant ssh frontend` to test out stuff manually on the command line.