FEATURE: Refactor Mellanox7800 firmware functions to be more robust
INTERNAL:
This completely refactors the parsing functions to scream loudly if
something about the format of the CLI responses being parsed changes.
The thought is that firmware upgrade/downgrade is important, and we
should scream loudly and fail quickly rather than botching a firmware
operation.

Many unit tests were added to confirm functionality and cover all
statements and branches of the refactored code.