BUGFIX: Wait for the quit command to run before terminating the process
INTERNAL:
This fixes a race condition where sendline() followed by terminate()
wouldn't guarantee that the command sent with sendline() would actually
be run on the target before terminating the process.

This also allows for all the arguments that are allowed in a say() call
to be passed as well, giving greater flexibility.