INTERNAL: Refactor system test-suite to a single extendable instance
There is now only a single system test-suite. The `--system-quick` flag to `run-tests.sh` is now just an alias for `--system` and will be removed once these changes make it to the `master` branch.

To test other pallets using the system test-suite, you would check out a copy of test-framework (from stacki repo), add a script to the `set-up.d` folder under the system test-suite, then call `run-tests.sh --system` passing a Stacki ISO to use as the frontend, and the `--extra-isos=` flag with the path to your pallet under test.

The `set-up.d` script should run all the commands on the frontend to add and enable the pallet(s) and then `stack enable discovery` with the appropriate arguments to install the pallets on the two backend nodes.

After the `set-up.d` scripts are ran, the two backends are powered up and will go through the provisioning process. Then pytest will be ran inside of the frontend VM, running all tests in the `tests` folder under `system`. You can place new tesst files or replace the existing files in that folder to customize the tests run. You may also place files in the `test-files` folder to make them available inside the frontend VM.