INTERNAL: Initial add of firmware version regex table
Implementation of this is incomplete (similar to the firmware_imp table)
and will have to be re-visited to integrate completely.

Thought was to require a regex pattern that would be used to both
	1. Validate the user supplied version number for a firmware file
	   makes sense.
	2. Process the version number returned from the target hardware.

This would help keep the version numbers honest so that version number
comparisons can be done with some modicum of faith that we're not
comparing apples to oranges.