INTERNAL: Only revert filesystem changes for /export/stack when needed
It turns out that out tests only care about modifications to /export/stack, and only for tests that do `add pallet` or `add cart`. So, these changes create a new fixture to only capture changes and revert /export/stack, which is used by tests that actually write to that directory.