FEATURE: Run test-framework integration tests in parallel
The tests are distributed across 4 cpus inside of the VM. Minor special handling for database names needed to be added to the stack command code to point everything at a copy of the database the test-framework manages. Each test file is pinned on the same process and it's test functions ran sequentially, which seems to keep things happy.

A new fixture named `exclusive_lock` has been added to allow your test to run with exclusive control of the VM. It does this by waiting for the other test processes to finish their tests, then pause them before they can start the next one. Only a few tests and fixtures that use shared resources (modify the filesystem, start servers on ports, inject code, etc) have needed this extreme step. Only use it if your test won't pass without it.

A second new fixture, named `debug_log`, has also been added. You add it to your other fixtures or tests and then call as a function. You give it a string and it will write it to a `debug.log` file in the reports directory, along with the test making the call and the xdist worker it is running on. Useful for debugging pytest fixtures because pytesst swallows normal print calls. You should remove any `debug_log` calls from your code before merging into `develop`, because writing to the shared log isn't very efficient.

I added a new `--audit` flag to the `run-tests.sh` script for the integration test suite. This will log any modifications to the `/etc`, `/export/stack/pallets`, and `/export/stack/carts` directories, and warn you at the end of testing if any of the tests are missing the fixtures to revery the filesystem modifications.

INTERNAL: Fix test_wsapi.py tests to not rely on side-effects of previous tests.

This entire test file needs to be refactored to test individual stack commands in isolation. The refactored tests should live in the `verb` folder for the command.