INTERNAL: A couple tweaks that might help recover from weird test hangs
This adds a little more delay between retrying a failing test. It also adds a timeout for individual tests, so we don't hang infinitely for some random reason.

I also gave the wsclient tests some love, pointing one of them to a testing Django server so it will hit our duplicate database, and renamed the other to reflect what it was testing.