BUGFIX: Add leading zeros to ID of script section
When post sections are generated, boot-post sections are generated
according to Stacki ordering, but are run in a lexical order. This
means that boot-post-20 is run after boot-post-100. If there was
an explicit ordering specified in Stacki, this ordering can get
ignored when AutoYaST actually runs.

Fix this by making the script ID a 4-digit ID with leading zeros.
'Cuz we'll never have more than 9999 sections in a node file. /flw