BUGFIX: Make enable pallet error message more explicit
When a pallet is trying to be enabled after being added, if a parameter mismatches,
the stack command would just say the pallet isn't valid. This isn't a clear enough
error message to the user so instead the error message was made more explicit. Now
when the same thing occurs, the error message includes a list of the parameters entered.