INTERNAL: We need more time to fetch the source.
Note: For SLES 11 system tests, I had to disable the check ran on the frontend which checks that the backends are reporting they are up. Once we figure out what is going wrong, we should enable them again.

There is also a bug in Jenkins where retries won't work when a timeout is in them, unless you catch the exception raised and then flag an error.