INTERNAL: Add report system test for detecting other DHCP servers
A problem that could occur since Stacki uses pxe boot to kick off the
booting process of backends is that another DHCP server other than
the frontend might get to the backend first and cause it to never
install. A way to detect this would be to send out a DHCP discover
packet from the frontend and if we get a response back, it probably
means there is a "rogue" DHCP server on a network it shouldn't be on
for Stacki.