BUGFIX: Fix permission error when installing host with new interface
When a backend is reinstalled and there is an interface not in the database,
the apache user would attempt to do a sync config and cause a permission error
to appear in the install logs. This isn't necessary during this stage (it's
worked fine until now with that error message) so pass in a flag to skip it.