INTERNAL: Fix issue of reload not handling multiple confirm prompts
This appears to happen when the switch thinks it is going to boot from
an older switch firmware image, which it appears to incorrectly think is
happening when upgrading from software 3.0.0.94 to 3.0.1.1.

This also required a change to expectmore to allow the use of
pexpect.spawn.send instead of pexpect.spawn.sendline. The x1052 switch
appears to treat the os.linesep that pexpect sends when using sendline
as auto cancelling the second confirmation prompt when one follows right
after the other.