BUGFIX: If an interface 'name' exists, then add it to /etc/hosts
INTERNAL: Remove references to `1.1.1.1` and `2.2.2.2` from tests.

BUGFIX: Remove `myhostname` from nsswitch.conf on Redhat.

It is unneeded because we always add localhost and localhost.localdomain to the /etc/hosts file we generate. Removing it brings Redhat inline with SLES, and prevents Redhat from resolving non-existant `.localdomain` subdomains to localhost.