FEATURE: `/etc/resolv.conf` improved handling
`resolv.search` attribute can override the search line.

Only the Frontend has access to the `Kickstart_PublicDNSServers`. For
all other nodes the Frontend will be the first nameserver IFF it is
serving DNS for any network (this was previously the case as well),
but the subsequent nameserver lines will come only from the
`Kickstart_PrivateDNSServers`.