INTERNAL: Blacklist report system on the web service
When trying to get the output of report system to show when running from the web
service, it was found that unless report system could run as root, tests would fail.
Since this could be a security risk and report system is not useful if some tests
always fail, it was decided it's best to blacklist it from the web service for now.