BUGFIX: Add command output to web server when command error is raised
This fixes a bug that cropped up when running report system on the web service,
as it was noticed only an output that a test failed was coming back to the client.
It was then discovered that only stderr would be given back as the web service
only returned it when a CommandError was raised. Now both stderr and stdout
are output. In addition if an exception is not raised for non sudo commands,
json output is checked for and if not it formats as json vs the previous behavior
of giving a stacktrace.

For the first fix if wsclient report system was run the output would be:

{"API Error": "error - One or more tests failed"}

Now with the fix:

{"API Error": "error - One or more tests failed", "Output":"$TESTOUTPUT"}

where $TESTOUTPUT is the results from report system

For the second fix, assuming a command doesn't raise an exception and does not
output valid json.

Before:

"$TESTOUTPUT"

Now:

{"Output": "$TESTOUTPUT"}