FEATURE: Prettier /etc/hosts files
Group by network not by host.

BUGFIX: /etc/hosts missing some entries

If host had a single interface but it was not marked as default
previously no entry was generate. Now assume the single interface is
the default.

If an interface was on a network w/o a zone no entry was generated.