INTERNAL: Add libvirt rpms and python library packages
This adds the libvirt-python package to foundation-python to allow
for Stacki to be able to communicate with kvm hypervisors. Libvirt-devel
is needed only in the bootstrap enviroment as it is only needed to build
the libvirt-python package, not at runtime. In addition libvirt-python is only needed
on the frontend so the manifest check for SLES 11 is excluded as it will never
be built there.