BUGFIX: Give self-signed certificates a unique serial
Browsers don't like it when your SSL cert has the same serial number (in this case 0) as a previous SSL cert for the same domain. This bits you when you are using `localhost` as your domain and you reinstall the frontend.

INTERNAL: Generate the RSA key for the self-signed cert using 4096 bits