INTERNAL: Add testing of backend partitioning to system tests
This uses the reworked report system test that checks backend
partitioning. This currently is only set up for redhat as for some
reason the example partitioning schemes are not working on SLES 12 KVM
hosts. Yes, it works on SLES 11 and SLES 12 if you use virtualbox. No,
I don't know why.