BUGFIX: Correct redhat os name for sync host firewall
This fixes a bug in sync host firewall on redhat based hosts.
Before the condition for calling iptables expected the string rhel
instead of redhat, with the latter being set as the os name
on redhat hosts. This resulted in a service that is sles only
trying to be called (stacki-iptables) and failing when trying to sync
the firewall.