FEATURE: Ship frontend-install.py as an RPM
INTERNAL:

The workflow for barnacling a frontend will now be to install the
foundation-python RPM and the stacki-fab RPM (which requires
foundation-python) in order to lay down frontend-install.py in
/opt/stack/bin.

Updated our various tools to use this new barnacle approach.

Changed the foundation-python RPM such that it now sets up its ldconf
file on install and runs ldconfig.

Update frontend-install.py now that it no longer has to be python 2
compatible, since the RPM requires foundation-python which we control.