INTERNAL: Add interface to kvm api in pylib
This commit adds a pylib library to interface with a kvm based hypervisor
for performing a variety of actions related to virtual machine deployment.
Once a connection to the kvm host is made (which requires ssh access),
virtual machine's can be added, removed, and modified via the library functions.
In addition, storage pool and volume creation/destruction is handled as well.
Also added is two new tables in the database schema for keeping track of virtual machine
and virtual machine disk information.