FEATURE: Add commands for syncing and generating config files for virtual machines
This commit adds a report and sync command for interfacing with hypervisors.
The report vm command allows for the generation of libvirt virtual machine configuration files
based on the information defined in Stacki. Sync vm is responsible for defining VM's,
creating/transferring storage, and updating the database after VM's or their storage has been removed.