INTERNAL: Pin ansible and all its dependencies because its a snowflake
Turns out that ansible doesn't have a sane setup.py, doesn't
specify its requirements in `install_requires`, and thus poetry cannot
figure out what its dependencies are. This forces us to manually pin
ansible and all its dependencies in the pyproject.toml file.

This also updates poetry to 1.0.0, fixing some breakages in pip2src that
came as part of that update.

This is a cherry-pick of the same commit from develop. Conflicts in
poetry.lock as well as LICENSE.txt had to be addressed.