INTERNAL: Update django to the latest version
This removes the symlinking of pymysql to mysqldb in favor of using the
more official workaround wherein pymysql installs itself as mysqldb in
the django settings file.

The pymysql version was bumped as part of this change.

This also fixes up some code using deprecated ways of doing things that
were removed in Django 2.x+.