FEATURE: Add package repositories as first-class objects includable in a stacki 'box'
Adds a repo arg processor, a yum template, and stack CRUD commands.  Changes the output of `stack list box`, to include the new repos column and adds a `stack list box repo` command.  `stack report host repo` was largely re-written to take advantage of the yum repo template and to merge the sles and redhat implementations into a yum implementation.

Note also, repo fields themselves can accept jinja variables referring to stacki attributes.  For example:

```
frontend-0-0:~ # stack list repo
NAME  ALIAS URL
test  test  http://{{ Kickstart_PrivateAddress }}/test/

frontend-0-0:~ # stack list host attr attr=Kickstart_PrivateAddress
HOST         SCOPE  TYPE  ATTR                     VALUE
frontend-0-0 global const Kickstart_PrivateAddress 192.168.0.2

```

In the next commit pallets and carts will use the same repo templating codepath and functionality, but currently hardcode the url to use `Kickstart_PrivateAddress`.  A future commit could allow replacing 'Kickstart_PrivateAddress' with something more flexible, allowing any given host to use a different ip to get to the Frontend's (or really, any server's) pallets.

I haven't extensively tested it, but `report host repo a:backend` took about half as much time (3s to 1.5s) for 100 backends.  Most of this performance boost probably comes from doing all the box data lookups and passing those to the implementation for each host, instead of doing the lookups for each host.  Test coverage is just below 100%.

An un-documented backdoor'd feature was included here as well.  A repo *may* be tied to a pallet.  In the event that it is undesirable to add a pallet in a box, but it contains repo(s) we want to provide, you may create a new repo, whose url points to the pallet web directory on the frontend and assign that to a box.  If the pallet is removed, the repo goes with it.  This was mostly intended for stacki develops to use in a pallet_hook - if for some reason the sles15 packages iso is not useable as is, or it is undesirable to add all repositories on it, we can create a pallet hook which adds as repo's the handful of repo's we care about.