INTERNAL: Provide more information when sub-commands fail
This ensures that the sub-command that was run is captured as most
CommandErrors assume the command in question is being run directly,
thus losing some context when being called from another command.

This provides a flag that defaults to on so commands can selectively
disable this feature. This also updates the scoped commands to use this
flag to turn this feature off, as they secretly use an internal only
scope parameter to delegate to a common command that holds all of the
logic.