INTERNAL: Fix periodic timeouts during CentOS tests
We have to run dhcp on eth0 for Vagrant, which under CentOS still causes dhclient-script to overwrite /etc/resolv.conf when we don't want to. We need to use the Vagrant host's resolv.conf so that it isn't pointed to libvirts' dnsmasq server running on the host, which resolves hostnames for our VM environment which shouldn't exist (backend-0-0 created by system test-suite, but shouldn't resolve in integration test-suite).

INTERNAL: Clean up some Ansible things to make output cleaner