FEATURE: Add hooks into pallet operations on the frontend
This allows for pallets to add scripts into a pallet_hooks folder at
the pallet root that will be run as specified in the
pallet_hooks/pallet_hooks.json.

This is mainly to enable SLES pallet patching and dropping the right
stacki initrd and vmlinuz on disk when supporting multiple versions of
the same OS, but this is a generic facility that pallets can use to
perform additional work during a `stack * pallet` operation.