INTERNAL: Add locking in the Vagrantfile used by test-framework
This change adds an interprocess lock via a Vagrant trigger to ensure only a single instance of `test-framework` is trying to grab a virtual bridge or a DHCP IP at a given time. The lock only lasts for 60 seconds (I had no way to easily hook Vagrant for when the network is set up), but that seems to be twice as much time as needed.

This change requires Vagrant version 2.2.0 (or newer), which has been released for over 1.5 years so should be a reasonable requirement.

Also, now that there is locking involved, I couldn't rely on simple sleeps to bring the backends up in a deterministic order. So, I wrote a little monitoring script to wait for one backend to come up before launching the next.