FEATURE: Initial support for IMAGES
A host should have either a BOX, IMAGE, or NFSROOT. This is just the
start to get some plumbing into stacki to make IMAGES first class
objects.

The prototyping for all of this was done on a virtualbox frontend
connected to a bunch of Pi4 boards which have been set to PXE boot
(pi4 does this in firmware). This is where the NFSROOT stuff comes in,
it is all entirely raspbian based (backend only). The goal is to have
a cluster of Pis that can the load an image and blast it to the sd
card, reboot, and then ask for a stacki machine profile.

Ignore the Pi stuff for now (need to write up how this all
works). This is really about getting started supporting images, and
potentially getting easy support for boutique flavors of Linux.

Changed Appliances to have the MANAGED and NODE (now called SUX, NODE
is deprecated) attributes into the table and computed as
attributes. With boxes becoming optional this made more sense, also
nuked the PUBLIC column (not used anymore).

The OS is still calculated (even though it shows up in `list host`),
the OS is determined (in order) from the BOX, IMAGE, or NFSROOT. No
effort is made to make sure these match if multiple are set.

- the database schema changed a lot
- probably broke a dozen tests

tldr - had a bunch of pis sitting around doing nothing and now I have
a cluster.